requirements.txt
streamlit
pandas
pycti
pandas
fpdf2